pyproject.toml
[project]
name = "store"
version = "0.1.0"
description = ""
authors = [
    {name = "Evgen Sumskyi",email = "grendersq12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
